fix(auth): CurrentUser -> CurrentUserEmail 로 번경

+++ src/main/java/com/entry/dexam/global/security/oauth/changer/TokenExchangerController.java
@@ -1,5 +1,6 @@
 package com.entry.dexam.global.security.oauth.changer;
 
+import com.entry.dexam.global.exception.exceptions.TokenNotFoundException;
 import com.entry.dexam.global.security.oauth.changer.dto.TokenExchangeResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,6 +20,9 @@ public class TokenExchangerController {
             @RequestParam(value="code") String code
     ) {
         ExchangeToken exchangeToken = exchangeTokenRedisRepository.findByCode(code);
+        if (exchangeToken == null) {
+            throw TokenNotFoundException.EXCEPTION;
+        }
         exchangeTokenRedisRepository.delete(exchangeToken);
 		return TokenExchangeResponse.builder().accessToken(
                 exchangeToken.getAccessToken()
